# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Industrial Lathe Tools API",
    "description": "Industrial Lathe Tools API - Complete product management system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/blog/{slug}": {
      "get": {
        "tags": [
          "Storefront",
          "Storefront"
        ],
        "summary": "Get Article",
        "operationId": "get_article_api_v1_blog__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/articles/{slug}": {
      "get": {
        "tags": [
          "Storefront",
          "Storefront"
        ],
        "summary": "Get Article Alias",
        "operationId": "get_article_alias_api_v1_articles__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contact": {
      "post": {
        "tags": [
          "Storefront",
          "Storefront"
        ],
        "summary": "Contact Us",
        "operationId": "contact_us_api_v1_contact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlogPostResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "cover_image": {
            "type": "string",
            "title": "Cover Image"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "reading_minutes": {
            "type": "integer",
            "title": "Reading Minutes"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "related_product_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Related Product Ids"
          },
          "blocks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Blocks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "cover_image",
          "published_at",
          "reading_minutes",
          "author"
        ],
        "title": "BlogPostResponse"
      },
...
      "ContactRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Full Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "subject": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Subject"
          },
          "message": {
            "type": "string",
            "minLength": 10,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "phone",
          "subject",
          "message"
        ],
        "title": "ContactRequest"
      },
      "ContactResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "ticket": {
            "type": "string",
            "title": "Ticket"
          }
        },
        "type": "object",
        "required": [
          "ticket"
        ],
        "title": "ContactResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}